improve search results component
